Remplissage des repo
- Possibilité de créer des repo avec ou sans log
- Affiche qqc

+++ Isen.DotNet.Projet.Library/Models/Implementation/Adresse.cs
@@ -8,8 +8,8 @@ namespace Isen.DotNet.Projet.Library.Models.Implementation
         public string Texte { get; set; }
         public long CodePostal { get; set; }
         public Commune Commune { get; set; }
-        public float Lattitude { get; set; }
-        public float Longitude { get; set; }
+        public double Lattitude { get; set; }
+        public double Longitude { get; set; }
     }
 
 }
