+++ Isen.DotNet.Projet.Library/Models/Implementation/Commune.cs
@@ -6,8 +6,8 @@ namespace Isen.DotNet.Projet.Library.Models.Implementation
     public class Commune : BaseModel
     {
         public string Nom { get; set; }
-        public float Lattitude { get; set; }
-        public float Longitude { get; set; }
+        public double Lattitude { get; set; }
+        public double Longitude { get; set; }
     }
 
 }
